Project Ugly: Converting to "long" moves instead of "int".

+++ src/main/java/com/debabrata/spotchess/support/notation/move/MoveParser.java
@@ -3,7 +3,7 @@ package com.debabrata.spotchess.support.notation.move;
 import com.debabrata.spotchess.types.Position;
 
 public interface MoveParser {
-    int getMove(Position position, String notation);
-    String getNotation(Position position, int move);
+    long getMove(Position position, String notation);
+    String getNotation(Position position, long move);
     boolean confirmFormat(String moveNotation);
 }
